requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.8
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
basicsr==1.4.2
blessed==1.20.0
blinker==1.5
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet==5.1.0
charset-normalizer==3.1.0
chumpy @ git+https://github.com/scottandrews/chumpy.git@fe51783e0364bf1e9b705541e7d77f894dd2b1ac
click==8.1.3
clip @ file:///home/yixuan/zyx/stablediffusion/CLIP
cmake==3.26.0
colorama==0.4.6
coloredlogs==15.0.1
configer==1.3.1
configparser==6.0.0
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
diffusers==0.12.1
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.0
entrypoints==0.4
filelock==3.10.0
fire==0.5.0
flatbuffers==23.3.3
flit-core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.43.1
freetype-py==2.4.0
frozenlist==1.3.3
fsspec==2023.9.2
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.16.2
google-auth-oauthlib==1.0.0
gpustat==1.1
grpcio==1.51.3
huggingface-hub==0.17.3
human-body-prior==0.8.5.0
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.5
imageio-ffmpeg==0.4.2
importlib-metadata==6.8.0
importlib-resources==5.12.0
invisible-watermark==0.1.5
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
jsonschema==4.17.3
kaggle==1.5.13
kiwisolver==1.4.5
latent-diffusion @ file:///home/yixuan/zyx/EasyCodeBase/models/latent_diffusion
lazy-loader==0.3
lit==15.0.7
lmdb==1.4.1
loguru==0.7.0
lora-diffusion @ git+https://github.com/cloneofsimo/lora.git@bdd51b04c49fa90a88919a19850ec3b4cf3c5ecd
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.3
mdurl==0.1.2
mediapipe==0.10.7
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mmcv==2.0.1
mmengine==0.9.0
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.13.1
onnxruntime==1.14.1
open-clip-torch @ file:///home/yixuan/zyx/HW/Diffusion-Prompt/third_party/open-clip-wheels/open_clip_wheels/open_clip_torch-2.14.0-py3-none-any.whl
opencv-contrib-python==4.8.1.78
opencv-python==4.7.0.72
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.8.1
protobuf==3.20.3
psutil==5.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.18.0
pydeck==0.8.0
pyDeprecate==0.3.1
pyglet==2.0.9
Pygments==2.15.1
PyOpenGL==3.1.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.1
pyrender==0.1.45
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==1.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
roma==1.4.1
rsa==4.9
sacremoses==0.0.53
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.32.0
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post5
smmap==5.0.0
smplpytorch==0.0.8
smplx==0.1.28
sounddevice==0.4.6
streamlit==0.73.1
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
taming-transformers @ file:///home/yixuan/zyx/EasyCodeBase/models/latent_diffusion/taming-transformers
tb-nightly==2.14.0a20230808
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
test-tube==0.7.5
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.6.13
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torch-fidelity==0.3.0
torchaudio==2.0.1
torchgeometry==0.1.2
torchmetrics==0.5.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
transformers==4.34.0
transforms3d==0.4.1
trimesh==4.0.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wandb==0.15.12
watchdog==2.3.1
wcwidth==0.2.6
Werkzeug==2.2.3
yapf==0.40.1
yarl==1.8.2
zipp==3.15.0
